Refactor GitLab API requests.
