+++ __init__.py
@@ -1,0 +1,3 @@
+"""
+Python script for interacting with the Thunderdome API.
+"""
