+++ fetch/__init__.py
@@ -1,0 +1,3 @@
+"""
+Functionality and data for the 'fetch' subcommand.
+"""
