+++ main.py
@@ -7,37 +7,13 @@ Main entrypoint for the application.
 import argparse
 import logging
 import re
-import typing
 
 import requests
 
-GITLAB_ORGA_MILESTONE_REGEX = re.compile(
-    r"https:\/\/gitlab\.com\/groups/(?P<orga>[a-zA-Z0-9\-\_]+)\/-\/"
-    r"milestones\/(?P<milestone>[0-9]+)"
-)
-GITLAB_PROJECT_MILESTONE_REGEX = re.compile(
-    r"https:\/\/gitlab\.com\/(?P<orga>[a-zA-Z0-9\-\_]+)\/(?:(?P<subgroup>[a-zA-Z0-9\-\_]+)\/)*"
-    r"(?P<project>[a-zA-Z0-9\-\_]+)\/-\/milestones\/(?P<milestone>[0-9]+)"
-)
-GITLAB_ITERATION_REGEX = re.compile(
-    r"https:\/\/gitlab\.com\/groups/(?P<orga>[a-zA-Z0-9\-\_]+)\/-\/"
-    r"cadences\/(?P<cadence>[0-9]+)\/iterations\/(?P<iteration>[0-9]+)"
-)
-GITLAB_EPIC_URL_REGEX = re.compile(
-    r"https:\/\/gitlab\.com\/groups\/(?P<orga>[a-zA-Z0-9\-\_]+)\/-\/"
-    r"epics\/(?P<epic>[0-9]+)"
-)
-GITLAB_PROJECT_URL_REGEX = re.compile(
-    r"https:\/\/gitlab\.com\/(?P<orga>[a-zA-Z0-9\-\_]+)\/(?:(?P<subgroup>[a-zA-Z0-9\-\_]+)\/)*"
-    r"(?P<project>[a-zA-Z0-9\-\_]+)"
-)
-GITLAB_ISSUE_URL_REGEX = re.compile(
-    r"https:\/\/gitlab\.com\/(?P<orga>[a-zA-Z0-9\-\_]+)\/(?:(?P<subgroup>[a-zA-Z0-9\-\_]+)\/)*"
-    r"(?P<project>[a-zA-Z0-9\-\_]+)\/-\/issues\/(?P<issue>[0-9]+)"
-)
-
-# Maximum number of issues to fetch per request
-GITLAB_PAGINATION_LIMIT = 100
+from util.definitions import GITLAB_ISSUE_URL_REGEX
+from util.gitlab_id import get_project_id
+from util.gitlab_issue import get_issues_from_epics, get_issues_from_iterations, \
+      get_issues_from_milestones, get_issues_from_projects, get_issue_info
 
 
 def parse_args() -> argparse.Namespace:
@@ -165,8 +141,8 @@ def fetch_plans(battle_id: str, api_key: str) -> list[dict]:
 
 
 def transfer_points(
-    plans: list[dict], 
-    gitlab_token: str, 
+    plans: list[dict],
+    gitlab_token: str,
     overwrite: bool = False
 ) -> None:
     """
@@ -384,7 +360,7 @@ def create_plans(args: argparse.Namespace) -> list[dict]:
 
 
 def create_plans_from_issues(
-    links: dict[int, str], 
+    links: dict[int, str],
     token: str,
     with_weighted: bool = False,
     with_closed: bool = False
@@ -436,345 +412,6 @@ def create_plans_from_issues(
             plans.append(plan)
 
     return plans
-
-
-def get_issues_from_milestones(links: list[str], token: str) -> dict[int, str]:
-    """
-    Get issues from GitLab milestones.
-
-    :param links: GitLab milestones to create plans from.
-    :param token: Token for the GitLab API.
-    """
-    logging.info("Fetching milestones from GitLab...")
-
-    gitlab_headers = {
-        "PRIVATE-TOKEN": token,
-    }
-
-    issues: dict[int, str] = {}
-    for link in links:
-        # check if the link is a group milestone
-        match = re.match(GITLAB_ORGA_MILESTONE_REGEX, link)
-        if not match:
-            logging.error(
-                "Invalid URL '%s' does not match GitLab URL pattern '%s'",
-                link,
-                GITLAB_ORGA_MILESTONE_REGEX.pattern,
-            )
-            continue
-
-        group_name = match.group("orga")
-
-        # get group ID
-        group_id = get_group_id(group_name, token)
-
-        # get milestone ID
-        milestone_iid = match.group("milestone")
-        gitlab_response = requests.get(
-            f"https://gitlab.com/api/v4/groups/{group_id}/milestones",
-            timeout=10,
-            params={"iids": [milestone_iid]},
-            headers=gitlab_headers,
-        )
-
-        if not gitlab_response.ok:
-            logging.error("Failed to fetch milestone %s", milestone_iid)
-            continue
-
-        payload = gitlab_response.json()
-        milestone_name = payload[0]["title"]
-
-        for res in paginate_request(
-            "https://gitlab.com/api/v4/issues",
-            {
-                "milestone": milestone_name, 
-                "per_page": GITLAB_PAGINATION_LIMIT, 
-                "scope": "all",
-            },
-            gitlab_headers,
-        ):
-            payload = res.json()
-            for issue in payload:
-                issues.update({issue["id"]: issue["web_url"]})
-
-    return issues
-
-
-def get_issues_from_iterations(links: list[str], token: str) -> dict[int, str]:
-    """
-    Create Thunderdome plans from GitLab iterations.
-
-    :param links: GitLab iterations to create plans from.
-    :param token: Token for the GitLab API.
-    """
-    logging.info("Fetching iterations from GitLab...")
-
-    gitlab_headers = {
-        "PRIVATE-TOKEN": token,
-    }
-
-    issues: dict[int, str] = {}
-    for link in links:
-        match = re.match(GITLAB_ITERATION_REGEX, link)
-        if not match:
-            logging.error(
-                "Invalid URL '%s' does not match GitLab URL pattern '%s'",
-                link,
-                GITLAB_ITERATION_REGEX.pattern,
-            )
-            continue
-
-        iteration_id = match.group("iteration")
-
-        for res in paginate_request(
-            "https://gitlab.com/api/v4/issues",
-            {
-                "iteration_id": iteration_id, 
-                "per_page": GITLAB_PAGINATION_LIMIT,
-                "scope": "all",
-            },
-            gitlab_headers,
-        ):
-            payload = res.json()
-            for issue in payload:
-                issues.update({issue["id"]: issue["web_url"]})
-
-    return issues
-
-
-def get_issues_from_projects(links: list[str], token: str) -> dict[int, str]:
-    """
-    Create Thunderdome plans from GitLab projects.
-
-    :param links: GitLab projects to create plans from.
-    :param token: Token for the GitLab API.
-    """
-    logging.info("Fetching projects from GitLab...")
-
-    gitlab_headers = {
-        "PRIVATE-TOKEN": token,
-    }
-
-    issues: dict[int, str] = {}
-    for link in links:
-        match = re.match(GITLAB_PROJECT_URL_REGEX, link)
-        if not match:
-            logging.error(
-                "Invalid URL '%s' does not match GitLab URL pattern '%s'",
-                link,
-                GITLAB_PROJECT_URL_REGEX.pattern,
-            )
-            continue
-
-        # get project ID
-        project_id = get_project_id(link, token, GITLAB_PROJECT_URL_REGEX)
-
-        for res in paginate_request(
-            f"https://gitlab.com/api/v4/projects/{project_id}/issues",
-            {
-                "per_page": GITLAB_PAGINATION_LIMIT, 
-                "scope": "all",
-            },
-            gitlab_headers,
-        ):
-            payload = res.json()
-            for issue in payload:
-                issues.update({issue["id"]: issue["web_url"]})
-
-    return issues
-
-
-def get_issues_from_epics(links: list[str], token: str) -> dict[int, str]:
-    """
-    Create Thunderdome plans from GitLab epics.
-
-    :param links: GitLab epics to create plans from.
-    :param token: Token for the GitLab API.
-    """
-    logging.info("Fetching epics from GitLab...")
-
-    gitlab_headers = {
-        "PRIVATE-TOKEN": token,
-    }
-
-    issues: dict[int, str] = {}
-    for link in links:
-        match = re.match(GITLAB_EPIC_URL_REGEX, link)
-        if not match:
-            logging.error(
-                "Invalid URL '%s' does not match GitLab URL pattern '%s'",
-                link,
-                GITLAB_EPIC_URL_REGEX.pattern,
-            )
-            continue
-
-        group_name = match.group("orga")
-        epic_iid = match.group("epic")
-
-        # get group ID
-        group_id = get_group_id(group_name, token)
-
-        for res in paginate_request(
-            f"https://gitlab.com/api/v4/groups/{group_id}/epics/{epic_iid}/issues",
-            {
-                "per_page": GITLAB_PAGINATION_LIMIT, 
-                "scope": "all",
-            },
-            gitlab_headers,
-        ):
-            payload = res.json()
-            for issue in payload:
-                issues.update({issue["id"]: issue["web_url"]})
-
-    return issues
-
-
-def get_issue_info(issue_link: str, token: str) -> dict | None:
-    """
-    Get information about a GitLab issue from the GitLab API.
-
-    :param link: Link to the GitLab issue.
-    :param token: Token for the GitLab API.
-    """
-    gitlab_headers = {
-        "PRIVATE-TOKEN": token,
-    }
-
-    match = re.match(GITLAB_ISSUE_URL_REGEX, issue_link)
-    if not match:
-        logging.error(
-            "Invalid URL '%s' does not match GitLab URL pattern '%s'",
-            issue_link,
-            GITLAB_ISSUE_URL_REGEX.pattern,
-        )
-        return None
-
-    project_path = match.group("project")
-    issue_iid = match.group("issue")
-
-    # Get project ID
-    project_id = get_project_id(issue_link, token, GITLAB_ISSUE_URL_REGEX)
-
-    # Get issue information
-    gitlab_response = requests.get(
-        f"https://gitlab.com/api/v4/projects/{project_id}/issues/{issue_iid}",
-        timeout=10,
-        headers=gitlab_headers,
-    )
-
-    if not gitlab_response.ok:
-        logging.error("Failed to fetch issue %s#%s", project_path, issue_iid)
-        return None
-
-    return gitlab_response.json()
-
-
-def get_group_id(group_path: str, token: str) -> int | None:
-    """
-    Get the ID of a GitLab group.
-
-    :param group_path: Group path name in the GitLab URL.
-    :param token: Token for the GitLab API.
-    """
-    gitlab_headers = {
-        "PRIVATE-TOKEN": token,
-    }
-
-    # Get group ID
-    gitlab_response = requests.get(
-        "https://gitlab.com/api/v4/groups",
-        timeout=10,
-        params={"search": group_path},
-        headers=gitlab_headers,
-    )
-
-    if not gitlab_response.ok:
-        logging.error("Failed to fetch group %s", group_path)
-        return None
-
-    payload = gitlab_response.json()
-    for group in payload:
-        if group["path"] == group_path:
-            group_id = group["id"]
-            logging.debug("Group %s has ID %s", group_path, group_id)
-            return group_id
-
-    logging.error("Failed to find ID of group %s", group_path)
-    return None
-
-
-def get_project_id(issue_link: str, token: str, regex) -> int | None:
-    """
-    Get the ID of a GitLab project from a GitLab issue.
-
-    :param issue_link: Link to the GitLab issue.
-    :param token: Token for the GitLab API.
-    """
-    gitlab_headers = {
-        "PRIVATE-TOKEN": token,
-    }
-
-    match = re.match(regex, issue_link)
-    if not match:
-        logging.error(
-            "Invalid URL '%s' does not match GitLab URL pattern '%s'",
-            issue_link,
-            regex.pattern,
-        )
-        return None
-
-    group_name = match.group("orga")
-    group_id = get_group_id(group_name, token)
-
-    project_path = match.group("project")
-
-    # Get project ID
-    gitlab_response = requests.get(
-        f"https://gitlab.com/api/v4/groups/{group_id}/search",
-        timeout=10,
-        params={"scope": "projects", "search": project_path},
-        headers=gitlab_headers,
-    )
-    payload = gitlab_response.json()
-
-    # Find project ID
-    for project in payload:
-        if project["path"] == project_path:
-            project_id = project["id"]
-            logging.debug("Project %s has ID %s", project_path, project_id)
-            return project_id
-
-    logging.error("Failed to find ID of project %s", project_path)
-    return None
-
-
-def paginate_request(
-    url: str, params: dict, headers: dict
-) -> typing.Generator[requests.Response, None, None]:
-    """
-    Paginate through a GitLab API request.
-
-    :param url: URL to the GitLab API.
-    :param params: Parameters for the request.
-    :param headers: Headers for the request.
-    :return: Response from the request.
-    """
-    response = requests.get(url, timeout=10, params=params, headers=headers)
-    if not response.ok:
-        logging.error("Failed to fetch %s", url)
-        return None
-
-    yield response
-
-    while "next" in response.links:
-        response = requests.get(
-            response.links["next"]["url"], timeout=10, headers=headers
-        )
-        if not response.ok:
-            logging.error("Failed to fetch %s", url)
-            return None
-
-        yield response
 
 
 if __name__ == "__main__":
