+++ util/__init__.py
@@ -1,0 +1,3 @@
+"""
+Shared functionality used in al subcommands.
+"""
